Add files via upload
Hey all, I did a little bit with the Liquor Sales file. I'm trying to use it for another class, and the file is huge. I did a tiny bit of work to subset it to 2015-2017. The resulting file is still huge. Feel free to fix it if the code needs work. We might want to use even fewer years, but we can decide that later.